fix not being dead when supposed to be bug
in the past, leaving ovl while u r supposed to be dead will revive
u client-side. this allows u to wander around while u appear to be
dead by the other players. not only does it make it very obvious
to the  other players that u r using some form of tool, this bug
also bears the potential to be used as a cheat as u r invincible in
that state and could run around collecting info about ur enemy
where abouts.