pubovl, add externalovl
"externalovl" is a tool that lets u or another player spectate while
not actually joined to the game.
"limbo" is the so called state (at least in openspades) where u r
connected to the server, but havent joined any team yet. its most
defining characteristic is the top-down perspective from which u
look at the map while the select team and weapon menu is shown.
using this command tricks ur or another player's client u want to use
it on into joining as a spectator only on his side.
since he hasnt spawned yet and doesnt show up on the scoreboard at
all, he is completely invisible to everyone else.
in that sense it is superior to the regular pubovl from this script.